Creating a proper Perl Module
@zeinebchelly we will need it later on when we want to integrate stuff. Looks complicated, but it's just a matter of applying a template for module creation using Module::Starter::PBP (this latter for Perl Best Practices)
This is basically needed because I saw I needed Switch.pm for running the module, and this kind of thing is better explained using a standard configuration which in Perl means using the Makefile.PL file to express dependencies. Later on we'll blow up everything for the module but for the time being we'll just use Makefile.PL so that when installing people can use the standard lines of installation,
Perl Makefile.PL
make
make intall

or just `cpanm --install-deps` if they want to install dependencies. I'm not sure you really need `Switch` but for the time being let's stick to that; we're going to need to express dependencies anyway