Update SampleFuzzyUse.pl
I have changed the example to let the context be more clear. I have changed the capital letters to lowercases for the variable names as you stated in your previous comment and I have added the strict and warning statements. I have 0 errors. Now all is fine file with this module :)
As you said the parametrization of this module should be kept in a separate file. We can do this later once we decide what is our next step.